Updated expenses tab
Put delete button inside the table
